Created preliminary particle impact effects
Particle impact effect only works properly for raycasts, and likely the scriptable object that i made to control the particle systems behavior will have to be replaced with a prefab based object pooling system to eliminate the need for coroutines because the coroutine doesn't ever end when using projectiles (my suspicion is that the projectile once destroyed cannot end the coroutine, and thus it continues). I also plan to implement another object pool for projectiles.